+++ components/bridge/dao/api/src/main/java/com/flipperdevices/bridge/dao/api/model/FlipperKeyType.kt
@@ -1,10 +1,10 @@
 package com.flipperdevices.bridge.dao.api.model
 
-import com.flipperdevices.core.ui.res.R as DesignSystem
 import androidx.annotation.DrawableRes
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.Color
 import com.flipperdevices.core.ui.theme.LocalPallet
+import com.flipperdevices.core.ui.res.R as DesignSystem
 
 enum class FlipperKeyType(
     val humanReadableName: String,